API_wrapper/requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# OpenAI
openai==1.10.0

# Vector Database & Search
faiss-cpu==1.7.4
numpy==1.26.3

# Caching
redis==5.0.1
hiredis==2.3.2

# NLP
spacy==3.7.2
# Run: python -m spacy download en_core_web_sm
sentence-transformers>=5.2.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14
